pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task Manager"
authors = ["enimalojd <enderston8@gmail.com>"]
readme = "README.md"
packages = [
    {include = "task_manager"}
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
django-bootstrap5 = "^24.1"
dj-database-url = "^2.1.0"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
psycopg2-binary = "^2.9.9"
django-ninja = "^1.1.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
